requirements.txt
torch
safetensors
torchvision
pillow
numpy
opencv-python
pytorch-lightning
pyiqa
torchmetrics
HyperPyYAML
tqdm
PyYAML
diffusers>=0.25.0
wandb
scikit-learn